Customer Regristration
Edited default regristration page to handle user input for accounts.  user can attatche themselves to an account.

Checks accounts to see if the account name is there.  Regristration fails if the account is there.

Checks contacts to see if newly entered SSN has a duplicate, ends regristration if there are duplicates.